Moved CPULinux class to details namespace
This commit relocates CPULinux class from the global namespace to 'detail'. This separation was done in order to better adhere to organizing internal working codes under detail namespace, and adhering to good code practices. Additionally, this enhances code clarity by indicating the role of this class as an internal detail. The CPULinux class mainly gets CPU properties in Linux. While moving, there are no changes in the logic or implementation of the class methods."