Update copyright and refactor OS related classes
Change the copyright headers in all edited files and refactor the OS related classes to a new structure that better represents the actual program design. The OS class is now an interface that provides the necessary OS information without dealing with the specifics of each operating system, with this, I was able to implement OSWin and OSLinux that are platform-specific implementations and are more suited to handle the specifics of their respective operating systems. This change brings a more clear program structure and makes future additions and modifications related to operating systems easier, while maintaining the same functionality that was present.