Rename system-specific and base files for better clarity
To improve clarity and maintainability of the code, system-specific files have been renamed to reflect the system they pertain to. For instance, files in the `hwinfo/apple` directory have been renamed from `os.hpp` to `osApple.hpp`, `battery.hpp` to `batteryApple.hpp`, etc. Similarly, files in the `base` directory were renamed to include `Base` in their filename. This should make it easier to understand and track the role of each file.