Optimize WMIWrapper queries and format
Refactored the WMIWrapper calls in the hwinfo module to use a single static instance rather than creating a new one for each call. This improves the efficiency by reusing the existing connection. Also, updated/format the if statements to be single-lined by changing the parameters in .clang-format file. The modification in .clang-format file enables consistent coding style across the code base which makes the code easily readable.