Refactor WMIWrapper usage in CPUWin class
This commit refactors the use of WMIWrapper instance in the CPUWin class. Instead of creating a new instance of WMIWrapper in each method where it's used, we now use a single, shared static instance (s_wmi) across all the methods.

This change reduces unnecessary memory allocations and also makes the code easier to grasp, as the static variable s_wmi is defined at the class level. This commits aims to improve both performance and clarity of the code.