Remove redundant WMI query method
The method `query` in `WMIInterface.hpp` was removed because of redundancy while getting system information. This method was used to query WMI for various system details including free physical memory. It was replaced by a more streamlined method `queryValue`. Additionally, the commented out code in `getFreeMemory()` in `ramWin.hpp` was removed, for a cleaner, more efficient codebase.