Refactor OS class structure for OS info extraction
This commit introduces a significant refactor to the OS class structure in the hwinfo library. The main objective of these changes was to streamline the process of OS information extraction, especially in the cases of window platforms.

Now, a new method getOSInfo() is introduced in the hwinfo namespace, which takes care of OS information extraction and returns an OS object directly. As a result, usage has been made easier in the client side where user no longer needs to instantiate OS object.

In the OSBase class, the properties have been encapsulated in an OSItem structure, thus enhancing clarity and management of different OS attributes.

On the Windows side, OSWin constructor has been default initialized and a new function getAllOSs() is introduced to get all the respective OS details. This was done to focus on OS data gathering in a single function that would return a well-constructed object. It enhances YAGNI principle by only invoking detailed OS information extraction when required.